scrape text regression test
