+++ autogpt/main.py
@@ -12,7 +12,7 @@ from autogpt.configurator import create_config
 from autogpt.logs import logger
 from autogpt.memory import get_memory
 from autogpt.plugins import scan_plugins
-from autogpt.prompts.prompt import construct_main_ai_config
+from autogpt.prompts.prompt import DEFAULT_TRIGGERING_PROMPT, construct_main_ai_config
 from autogpt.utils import get_current_git_branch, get_latest_bulletin
 from autogpt.workspace import Workspace
 from scripts.install_plugin_deps import install_plugin_dependencies
@@ -124,11 +124,7 @@ def run_auto_gpt(
     # Initialize variables
     full_message_history = []
     next_action_count = 0
-    # Make a constant:
-    triggering_prompt = (
-        "Determine which next command to use, and respond using the"
-        " format specified above:"
-    )
+
     # Initialize memory and make sure it is empty.
     # this is particularly important for indexing and referencing pinecone memory
     memory = get_memory(cfg, init=True)
@@ -148,7 +144,7 @@ def run_auto_gpt(
         command_registry=command_registry,
         config=ai_config,
         system_prompt=system_prompt,
-        triggering_prompt=triggering_prompt,
+        triggering_prompt=DEFAULT_TRIGGERING_PROMPT,
         workspace_directory=workspace_directory,
     )
     agent.start_interaction_loop()
