+++ autogpt/prompts/prompt.py
@@ -10,6 +10,10 @@ from autogpt.utils import clean_input
 
 CFG = Config()
 
+DEFAULT_TRIGGERING_PROMPT = (
+    "Determine which next command to use, and respond using the format specified above:"
+)
+
 
 def build_default_prompt_generator() -> PromptGenerator:
     """
@@ -70,9 +74,6 @@ def build_default_prompt_generator() -> PromptGenerator:
     prompt_generator.add_performance_evaluation(
         "Every command has a cost, so be smart and efficient. Aim to complete tasks in"
         " the least number of steps."
-    )
-    prompt_generator.add_performance_evaluation(
-        "If you cannot think of a valid command to perform start or message an agent to determine the next command."
     )
     prompt_generator.add_performance_evaluation("Write all code to a file.")
     return prompt_generator
