+++ tests/integration/goal_oriented/test_write_file.py
@@ -1,14 +1,11 @@
-import concurrent
 import os
-import unittest
 
 import pytest
 
-from autogpt.agent import Agent
-from autogpt.commands.command import CommandRegistry
-from autogpt.commands.file_operations import delete_file, read_file
-from autogpt.config import AIConfig, Config
-from autogpt.memory import get_memory
+from autogpt.commands.file_operations import read_file
+from autogpt.config import Config
+from tests.integration.agent_factory import create_writer_agent
+from tests.integration.agent_utils import run_interaction_loop
 from tests.utils import requires_api_key
 
 CFG = Config()
@@ -23,59 +20,8 @@ def test_write_file(workspace) -> None:
     file_name = str(workspace.get_path("hello_world.txt"))
     agent = create_writer_agent(workspace)
     try:
-        with concurrent.futures.ThreadPoolExecutor() as executor:
-            future = executor.submit(agent.start_interaction_loop)
-            try:
-                result = future.result(timeout=45)
-            except concurrent.futures.TimeoutError:
-                assert False, "The process took longer than 45 seconds to complete."
+        run_interaction_loop(agent, 40)
     # catch system exit exceptions
     except SystemExit:  # the agent returns an exception when it shuts down
         content = read_file(file_name)
         assert content == "Hello World", f"Expected 'Hello World', got {content}"
-
-
-def create_writer_agent(workspace):
-    command_registry = CommandRegistry()
-    command_registry.import_commands("autogpt.commands.file_operations")
-    command_registry.import_commands("autogpt.app")
-
-    ai_config = AIConfig(
-        ai_name="write_to_file-GPT",
-        ai_role="an AI designed to use the write_to_file command to write 'Hello World' into a file named \"hello_world.txt\" and then use the task_complete command to complete the task.",
-        ai_goals=[
-            "Use the write_to_file command to write 'Hello World' into a file named \"hello_world.txt\".",
-            "Use the task_complete command to complete the task.",
-            "Do not use any other commands.",
-        ],
-    )
-    ai_config.command_registry = command_registry
-    CFG.set_continuous_mode(True)
-    CFG.set_memory_backend("no_memory")
-    CFG.set_temperature(0)
-    memory = get_memory(CFG, init=True)
-    triggering_prompt = (
-        "Determine which next command to use, and respond using the"
-        " format specified above:"
-    )
-    system_prompt = ai_config.construct_full_prompt()
-
-    agent = Agent(
-        ai_name="",
-        memory=memory,
-        full_message_history=[],
-        command_registry=command_registry,
-        config=ai_config,
-        next_action_count=0,
-        system_prompt=system_prompt,
-        triggering_prompt=triggering_prompt,
-        workspace_directory=workspace.root,
-    )
-
-    os.environ["TIKTOKEN_CACHE_DIR"] = ""
-
-    return agent
-
-
-if __name__ == "__main__":
-    unittest.main()
