+++ tests/utils.py
@@ -40,3 +40,7 @@ def skip_in_ci(test_function):
         os.environ.get("CI") == "true",
         reason="This test doesn't work on GitHub Actions.",
     )(test_function)
+
+
+def get_workspace_file_path(workspace, file_name):
+    return str(workspace.get_path(file_name))
